move endpoint counter to its own file
